Fixed deprecation warning that was caused by 'self.nrOfDetectionAngleSteps' being of type np.float64. This caused the warning below. Changing it to np.int32 also caused the warning. Changing it to 'int'. Resolved the issue. This was the warning: VisibleDeprecationWarning: using a non-integer number instead of an integer will result in an error in the future

+++ TrackingAlgorithm/TrackingAlgorithm/contourTrackerMainClass.py
@@ -69,7 +69,7 @@ class contourTrackerMain( object ):
 		pass
 
 	def setupClVariables(self):
-		self.nrOfDetectionAngleSteps = self.detectionKernelStrideSize * self.nrOfStrides
+		self.nrOfDetectionAngleSteps = int(self.detectionKernelStrideSize * self.nrOfStrides)
 		self.host_mostRecentMembraneCoordinatesX = np.zeros(shape=self.nrOfDetectionAngleSteps,dtype=np.float64)
 		self.dev_mostRecentMembraneCoordinatesX = cl_array.to_device(self.managementQueue, self.host_mostRecentMembraneCoordinatesX)
 		self.host_mostRecentMembraneCoordinatesY = np.zeros(shape=self.nrOfDetectionAngleSteps,dtype=np.float64)
